fixing return type of size to int
git-svn-id: https://www.evllabs.com/svn/jgaap/trunk@248 bcaee279-2e94-4b2e-be96-a1d7c6f3fc26

+++ src/com/jgaap/generics/EventHistogram.java
@@ -164,7 +164,7 @@ public class EventHistogram implements Iterable<Event> {
 	 * 
 	 * @return the number of distinct Event types stored in a histogram
 	 */
-	public long getNTypes() {
+	public int getNTypes() {
 		return theHist.size();
 	}
